requirements.txt
flask==2.3.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.0
pytest-cov==4.1.0
